+++ src/models/index.ts
@@ -1,9 +1,7 @@
-import { createModel } from '@rematch/core';
-import Post from '../out/PostList.json';
-import Octokit from '@octokit/rest';
+import posts from './posts';
+import issuesAndPRs from './issueAndPR';
 
-// const octokit = new Octokit();
-// octokit.search.issuesAndPullRequests({
-//   q:
-// })
-// export const 
+export {
+  posts,
+  issuesAndPRs,
+};
